feat: Add new components for task management including CategoryPill, DateButton, PriorityButton, and RecurrencePickerSheet
