+++ lich-plus/Features/Tasks/Components/CreateItemSheet.swift
@@ -510,263 +510,6 @@ struct CreateItemSheet: View {
     }
 }
 
-// MARK: - Supporting Views
-
-struct FormSection<Content: View>: View {
-    let title: String
-    @ViewBuilder let content: Content
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: AppTheme.spacing8) {
-            Text(title)
-                .font(.system(size: AppTheme.fontBody, weight: .medium))
-                .foregroundStyle(AppColors.textSecondary)
-
-            content
-        }
-    }
-}
-
-struct DateButton: View {
-    let date: Date
-    let action: () -> Void
-
-    private var dateFormatter: DateFormatter {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "d MMM, h:mm a"
-        return formatter
-    }
-
-    var body: some View {
-        Button(action: action) {
-            Text(dateFormatter.string(from: date))
-                .font(.system(size: AppTheme.fontBody))
-                .foregroundStyle(AppColors.textPrimary)
-                .frame(maxWidth: .infinity)
-                .padding(.vertical, AppTheme.spacing12)
-                .background(AppColors.background)
-                .cornerRadius(AppTheme.cornerRadiusLarge)
-                .overlay(
-                    RoundedRectangle(cornerRadius: AppTheme.cornerRadiusLarge)
-                        .stroke(AppColors.borderLight, lineWidth: 1)
-                )
-        }
-        .buttonStyle(.plain)
-    }
-}
-
-struct CategoryPill: View {
-    let category: TaskCategory
-    let isSelected: Bool
-    let action: () -> Void
-
-    private var backgroundColor: Color {
-        switch category {
-        case .work:
-            return AppColors.categoryWorkBackground
-        case .personal:
-            return AppColors.categoryPersonalBackground
-        case .meeting:
-            return AppColors.categoryMeetingBackground
-        case .birthday:
-            return AppColors.categoryBirthdayBackground
-        case .holiday:
-            return AppColors.categoryHolidayBackground
-        case .other:
-            return AppColors.categoryOtherBackground
-        }
-    }
-
-    private var textColor: Color {
-        switch category {
-        case .work:
-            return AppColors.categoryWorkText
-        case .personal:
-            return AppColors.categoryPersonalText
-        case .meeting:
-            return AppColors.categoryMeetingText
-        case .birthday:
-            return AppColors.categoryBirthdayText
-        case .holiday:
-            return AppColors.categoryHolidayText
-        case .other:
-            return AppColors.categoryOtherText
-        }
-    }
-
-    private var iconName: String {
-        switch category {
-        case .work:
-            return "briefcase.fill"
-        case .personal:
-            return "heart.fill"
-        case .meeting:
-            return "person.2.fill"
-        case .birthday:
-            return "gift.fill"
-        case .holiday:
-            return "sun.max.fill"
-        case .other:
-            return "tag.fill"
-        }
-    }
-
-    var body: some View {
-        Button(action: action) {
-            HStack(spacing: AppTheme.spacing4) {
-                Image(systemName: iconName)
-                    .font(.system(size: 12))
-
-                Text(category.displayName)
-                    .font(.system(size: AppTheme.fontBody, weight: .medium))
-            }
-            .foregroundStyle(textColor)
-            .padding(.horizontal, AppTheme.spacing12)
-            .padding(.vertical, AppTheme.spacing8)
-            .background(backgroundColor)
-            .cornerRadius(AppTheme.cornerRadiusLarge)
-            .overlay(
-                RoundedRectangle(cornerRadius: AppTheme.cornerRadiusLarge)
-                    .stroke(isSelected ? textColor : Color.clear, lineWidth: 2)
-            )
-        }
-        .buttonStyle(.plain)
-    }
-}
-
-struct PriorityButton: View {
-    let priority: Priority
-    let isSelected: Bool
-    let action: () -> Void
-
-    private var borderColor: Color {
-        switch priority {
-        case .high:
-            return AppColors.primary
-        case .medium:
-            return AppColors.eventOrange
-        case .low, .none:
-            return AppColors.borderLight
-        }
-    }
-
-    private var textColor: Color {
-        switch priority {
-        case .high:
-            return AppColors.primary
-        case .medium:
-            return AppColors.eventOrange
-        case .low, .none:
-            return AppColors.textSecondary
-        }
-    }
-
-    private var backgroundColor: Color {
-        guard isSelected else { return AppColors.background }
-        switch priority {
-        case .high:
-            return AppColors.priorityHighBackground
-        case .medium:
-            return AppColors.priorityMediumBackground
-        case .low, .none:
-            return AppColors.background
-        }
-    }
-
-    var body: some View {
-        Button(action: action) {
-            Text(priority.displayName)
-                .font(.system(size: AppTheme.fontBody, weight: .medium))
-                .foregroundStyle(textColor)
-                .frame(maxWidth: .infinity)
-                .padding(.vertical, AppTheme.spacing12)
-                .background(backgroundColor)
-                .cornerRadius(AppTheme.cornerRadiusLarge)
-                .overlay(
-                    RoundedRectangle(cornerRadius: AppTheme.cornerRadiusLarge)
-                        .stroke(borderColor, lineWidth: isSelected ? 2 : 1)
-                )
-        }
-        .buttonStyle(.plain)
-    }
-}
-
-// MARK: - Date Picker Sheet
-
-struct DatePickerSheet: View {
-    let title: String
-    @Binding var selectedDate: Date
-    let onDone: () -> Void
-
-    var body: some View {
-        NavigationStack {
-            VStack {
-                DatePicker(
-                    "",
-                    selection: $selectedDate,
-                    displayedComponents: [.date, .hourAndMinute]
-                )
-                .datePickerStyle(.graphical)
-                .labelsHidden()
-                .padding()
-
-                Spacer()
-            }
-            .navigationTitle(title)
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                ToolbarItem(placement: .confirmationAction) {
-                    Button(String(localized: "task.done")) {
-                        onDone()
-                    }
-                    .foregroundStyle(AppColors.primary)
-                }
-            }
-        }
-    }
-}
-
-// MARK: - Recurrence Picker Sheet
-
-struct RecurrencePickerSheet: View {
-    @Binding var selectedRecurrence: RecurrenceType
-    let onDone: () -> Void
-
-    var body: some View {
-        NavigationStack {
-            List {
-                ForEach(RecurrenceType.allCases, id: \.self) { recurrence in
-                    Button {
-                        selectedRecurrence = recurrence
-                    } label: {
-                        HStack {
-                            Text(recurrence.displayName)
-                                .foregroundStyle(AppColors.textPrimary)
-
-                            Spacer()
-
-                            if selectedRecurrence == recurrence {
-                                Image(systemName: "checkmark")
-                                    .foregroundStyle(AppColors.primary)
-                            }
-                        }
-                    }
-                }
-            }
-            .navigationTitle(String(localized: "createItem.recurrence"))
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                ToolbarItem(placement: .confirmationAction) {
-                    Button(String(localized: "task.done")) {
-                        onDone()
-                    }
-                    .foregroundStyle(AppColors.primary)
-                }
-            }
-        }
-    }
-}
-
 // MARK: - Preview
 
 #Preview {
